+++ src/Microsoft.ApplicationInsights/Extensibility/Implementation/OperationHolder.cs
@@ -71,7 +71,7 @@
                             var currentActivity = Activity.Current;
                             if (currentActivity == null 
                             || (Activity.DefaultIdFormat != ActivityIdFormat.W3C && operationTelemetry.Id != currentActivity.Id) 
-                            || (Activity.DefaultIdFormat == ActivityIdFormat.W3C && operationTelemetry.Id != W3CActivityExtensions.FormatTelemetryId(currentActivity.TraceId.ToHexString(), currentActivity.SpanId.ToHexString()) ))
+                            || (Activity.DefaultIdFormat == ActivityIdFormat.W3C && operationTelemetry.Id != W3CUtilities.FormatTelemetryId(currentActivity.TraceId.ToHexString(), currentActivity.SpanId.ToHexString()) ))
                             {
                                 // W3COperationCorrelationTelemetryInitializer changes Id
                                 // but keeps an original one in 'ai_legacyRequestId' property
